+++ src/components/ImageCard/ImageHotspots.jsx
@@ -2,7 +2,7 @@ import React, { useMemo, useState } from 'react';
 import useDeepCompareEffect from 'use-deep-compare-effect';
 import PropTypes from 'prop-types';
 
-import { InlineLoading } from '../..';
+import { InlineLoading } from '../InlineLoading';
 
 import Hotspot, { propTypes as HotspotPropTypes } from './Hotspot';
 import ImageControls from './ImageControls';
