+++ src/components/InlineLoading/index.js
@@ -1,0 +1,1 @@
+export { InlineLoading } from 'carbon-components-react';
